Move my routes component to new dir
