+++ src/components/routes/MyRoutes.js
@@ -1,39 +1,39 @@
 import {Navigate, Route, Routes} from "react-router-dom";
 import React from "react";
-import SignupSuccess from "./pages/SignUpSuccess";
-import {ROUTES} from "../helpers/constants/configConstants";
-import {Home} from "./pages/Home";
-import {Dashboard} from "./pages/Dashboard";
-import {OverviewPane} from "./dashboard/dashboardTabPanes/Overview/OverviewPane";
-import {MyTitlesPane} from "./dashboard/dashboardTabPanes/MyTitles/MyTitlesPane";
-import {OtherCollectionsPane} from "./dashboard/dashboardTabPanes/OtherCollections/OtherCollectionsPane";
-import Profile from "./pages/Profile";
-import {Title} from "./pages/Title";
-import {Titles} from "./pages/Titles";
-import {useAppContext} from "../context/AppContext";
-import {Admin} from "./pages/admin/Admin";
-import {AdminPublisher} from "./pages/admin/publisher/AdminPublisher";
-import {AdminPublishers} from "./pages/admin/publisher/AdminPublishers";
-import {AdminPublisherAdd} from "./pages/admin/publisher/AdminPublisherAdd";
-import {AdminTitle} from "./pages/admin/title/AdminTitle";
-import {AdminTitles} from "./pages/admin/title/AdminTitles";
-import {AdminTitleAdd} from "./pages/admin/title/AdminTitleAdd";
-import {AdminIssue} from "./pages/admin/issue/AdminIssue";
-import {AdminIssues} from "./pages/admin/issue/AdminIssues";
-import {Issue} from "./pages/Issue";
-import {Issues} from "./pages/Issues";
-import {Publisher} from "./pages/Publisher";
-import {Publishers} from "./pages/Publishers";
-import {AdminUsers} from "./pages/admin/users/AdminUsers";
-import {Marvelklubben} from "./pages/Marvelklubben";
-import ChangePassword from "./pages/ChangePassword";
-import {User} from "./pages/User";
-import {Users} from "./pages/Users";
-import {AdminMessages} from "./pages/admin/message/AdminMessages";
-import {AdminMessage} from "./pages/admin/message/AdminMessage";
-import {NoMatch} from "./pages/NoMatch";
-import {GradeValues} from "./pages/GradeValues";
-import {ValuationPane} from "./dashboard/dashboardTabPanes/Valuation/ValuationPane";
+import SignupSuccess from "../pages/SignUpSuccess";
+import {ROUTES} from "../../helpers/constants/configConstants";
+import {Home} from "../pages/Home";
+import {Dashboard} from "../pages/Dashboard";
+import {OverviewPane} from "../dashboard/dashboardTabPanes/Overview/OverviewPane";
+import {MyTitlesPane} from "../dashboard/dashboardTabPanes/MyTitles/MyTitlesPane";
+import {OtherCollectionsPane} from "../dashboard/dashboardTabPanes/OtherCollections/OtherCollectionsPane";
+import Profile from "../pages/Profile";
+import {Title} from "../pages/Title";
+import {Titles} from "../pages/Titles";
+import {useAppContext} from "../../context/AppContext";
+import {Admin} from "../pages/admin/Admin";
+import {AdminPublisher} from "../pages/admin/publisher/AdminPublisher";
+import {AdminPublishers} from "../pages/admin/publisher/AdminPublishers";
+import {AdminPublisherAdd} from "../pages/admin/publisher/AdminPublisherAdd";
+import {AdminTitle} from "../pages/admin/title/AdminTitle";
+import {AdminTitles} from "../pages/admin/title/AdminTitles";
+import {AdminTitleAdd} from "../pages/admin/title/AdminTitleAdd";
+import {AdminIssue} from "../pages/admin/issue/AdminIssue";
+import {AdminIssues} from "../pages/admin/issue/AdminIssues";
+import {Issue} from "../pages/Issue";
+import {Issues} from "../pages/Issues";
+import {Publisher} from "../pages/Publisher";
+import {Publishers} from "../pages/Publishers";
+import {AdminUsers} from "../pages/admin/users/AdminUsers";
+import {Marvelklubben} from "../pages/Marvelklubben";
+import ChangePassword from "../pages/ChangePassword";
+import {User} from "../pages/User";
+import {Users} from "../pages/Users";
+import {AdminMessages} from "../pages/admin/message/AdminMessages";
+import {AdminMessage} from "../pages/admin/message/AdminMessage";
+import {NoMatch} from "../pages/NoMatch";
+import {GradeValues} from "../pages/GradeValues";
+import {ValuationPane} from "../dashboard/dashboardTabPanes/Valuation/ValuationPane";
 
 
 export const MyRoutes = () => {
